hydrolab/requirements.txt
smbus2
RPi.GPIO
picamera2
paho-mqtt
firebase-admin
python-dotenv
